msm_execute_helper: Fix -Wsign-conversion warning

+++ tools/msm_execute_helper.c
@@ -5,8 +5,8 @@
 
 int main(int argc, char* argv[])
 {
-    gchar **argv_modified = g_new0 (gchar *, argc + 1);
-    memcpy (argv_modified, argv, argc * sizeof (char*));
+    gchar **argv_modified = g_new0 (gchar *, (size_t)(argc + 1));
+    memcpy (argv_modified, argv, (size_t)(argc) * sizeof (char*));
     argv_modified[0] = COMMAND;
     int errsv = 0;
